+++ c7n/exceptions.py
@@ -12,6 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+try:
+    from botocore.exceptions import ClientError
+except ImportError:
+    class ClientError(Exception): pass
+
 
 class CustodianError(Exception):
     """Custodian Exception Base Class
