+++ c7n/log.py
@@ -24,8 +24,7 @@ uses a separate session.
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import boto3
-from botocore.exceptions import ClientError
+from c7n.exceptions import ClientError
 
 import itertools
 import logging
@@ -72,7 +71,7 @@ class CloudWatchLogHandler(logging.Handler):
         super(CloudWatchLogHandler, self).__init__()
         self.log_group = log_group
         self.log_stream = log_stream
-        self.session_factory = session_factory or boto3.Session
+        self.session_factory = session_factory
         self.transport = None
         self.queue = Queue.Queue()
         self.threads = []
