+++ c7n/utils.py
@@ -13,9 +13,6 @@
 # limitations under the License.
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from botocore.exceptions import ClientError
-
-import boto3
 import copy
 import csv
 from datetime import datetime, timedelta
@@ -32,6 +29,7 @@ import six
 import sys
 
 
+from c7n.exceptions import ClientError
 from c7n import ipaddress
 
 # Try to place nice in lambda exec environment
@@ -451,17 +449,36 @@ def reformat_schema(model):
     return ret
 
 
-_profile_session = None
+# from botocore.utils avoiding runtime dependency for botocore for other providers.
+# license apache 2.0
+def set_value_from_jmespath(source, expression, value, is_first=True):
+    # This takes a (limited) jmespath-like expression & can set a value based
+    # on it.
+    # Limitations:
+    # * Only handles dotted lookups
+    # * No offsets/wildcards/slices/etc.
+    bits = expression.split('.', 1)
+    current_key, remainder = bits[0], bits[1] if len(bits) > 1 else ''
 
+    if not current_key:
+        raise ValueError(expression)
 
-def get_profile_session(options):
-    global _profile_session
-    if _profile_session:
-        return _profile_session
+    if remainder:
+        if current_key not in source:
+            # We've got something in the expression that's not present in the
+            # source (new key). If there's any more bits, we'll set the key
+            # with an empty dictionary.
+            source[current_key] = {}
 
-    profile = getattr(options, 'profile', None)
-    _profile_session = boto3.Session(profile_name=profile)
-    return _profile_session
+        return set_value_from_jmespath(
+            source[current_key],
+            remainder,
+            value,
+            is_first=False
+        )
+
+    # If we're down to a single key, set it.
+    source[current_key] = value
 
 
 def format_string_values(obj, *args, **kwargs):
